requirements.txt
# FinOps Dashboard - Free Tier Requirements
# Minimal dependencies for AWS free tier cost monitoring

# Core AWS SDK (Free)
boto3==1.34.144
botocore==1.34.144

# Configuration and data handling (Free)
PyYAML==6.0.1
python-dotenv==1.0.1

# SQLite is built into Python (Free)
# No additional database dependencies needed

# Data visualization and reporting (Free alternatives to paid services)
matplotlib==3.8.4
plotly==5.19.0

# Email reporting (using free SMTP)
smtplib  # Built into Python

# Scheduling and automation
schedule==1.2.2

# Development and testing
pytest==8.2.2
pytest-cov==5.0.0

# Optional: For advanced analysis (all free)
pandas==2.2.2
numpy==1.26.4

# Optional: For web dashboard (lightweight alternative to Grafana)
flask==3.0.3
dash==2.17.1

# Logging and monitoring
structlog==24.1.0

# System utilities
psutil==5.9.8

# Date/time utilities
python-dateutil==2.9.0

# Note: All these packages are free and open source
# AWS Cost Explorer API is also free (within reasonable limits)
# SQLite database is free and lightweight
# This setup avoids paid services like:
# - Premium Grafana Cloud
# - Paid database services
# - Commercial monitoring tools
